requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.3.0
langchain-anthropic>=0.1.1
anthropic>=0.8.0
jinja2>=3.1.2
python-dotenv>=1.0.0

# Analysis and visualization dependencies
pandas>=2.0.0
matplotlib>=3.5.0
openpyxl>=3.0.9